Import live.2008.11.04.tar.gz
- Fixed a bug in "RTPInterface" that was sometimes causing sockets not to be reclaimed properly by a
  RTSP server when streaming RTP/RTCP-over-TCP.  (Thanks to Ralf Globisch for tracking this down.)
- Changed the RTSP 'start time' parameter from "float" to "double", to allow for
  accurate 'start times' that are very large - e.g., Unix 'epoch' times.
  (Thanks to Morgan Toervolt for the suggestion and patch.)
- Made some simplifications to the 'reception stats' code in "RTPSource".
  (Thanks to Guy Bonneau for this suggestion.)
- Fixed a minor bug in "QuickTimeFileSink". (Thanks to Pramod Bhagwat.)

+++ liveMedia/MP3AudioFileServerMediaSubsession.cpp
@@ -48,7 +48,7 @@ MP3AudioFileServerMediaSubsession
 }
 
 void MP3AudioFileServerMediaSubsession
-::seekStreamSource(FramedSource* inputSource, float seekNPT) {
+::seekStreamSource(FramedSource* inputSource, double seekNPT) {
   MP3FileSource* mp3Source;
   if (fGenerateADUs) {
     // "inputSource" is a filter; use its input source instead.
