+++ liveMedia/include/MPEG1or2DemuxedServerMediaSubsession.hh
@@ -45,7 +45,7 @@ private:
   virtual ~MPEG1or2DemuxedServerMediaSubsession();
 
 private: // redefined virtual functions
-  virtual void seekStreamSource(FramedSource* inputSource, float seekNPT);
+  virtual void seekStreamSource(FramedSource* inputSource, double seekNPT);
   virtual FramedSource* createNewStreamSource(unsigned clientSessionId,
 					      unsigned& estBitrate);
   virtual RTPSink* createNewRTPSink(Groupsock* rtpGroupsock,
