+++ liveMedia/include/MPEG2TransportFileServerMediaSubsession.hh
@@ -57,7 +57,7 @@ private: // redefined virtual functions
                            unsigned short& rtpSeqNum,
                            unsigned& rtpTimestamp);
   virtual void pauseStream(unsigned clientSessionId, void* streamToken);
-  virtual void seekStream(unsigned clientSessionId, void* streamToken, float seekNPT);
+  virtual void seekStream(unsigned clientSessionId, void* streamToken, double seekNPT);
   virtual void setStreamScale(unsigned clientSessionId, void* streamToken, float scale);
   virtual void deleteStream(unsigned clientSessionId, void*& streamToken);
